refactor to add multiple data sources
